don't need u32 lut if we don't use large sf

+++ plonky2/ml/src/gadgets/bias_div_round_relu6.rs
@@ -74,7 +74,7 @@ impl<F: RichField + Extendable<D>, const D: usize> Gadget<F, D> for BiasDivRound
         .map(|d| {
           let i = d as i64;
           let val = map.get(&i).unwrap();
-          (i as u32, *val as u32)
+          (i as u16, *val as u16)
         })
         .collect::<Vec<_>>()
     });
